requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlmodel==0.0.21
httpx==0.27.0
twilio==9.2.3
vonage==3.14.0
pydantic==2.9.2


